added all of the handler

+++ 42_Heilbronn/Basecamp/printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: jkollner <jkollner@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/08 12:38:32 by jkollner          #+#    #+#             */
-/*   Updated: 2022/11/10 14:41:28 by jkollner         ###   ########.fr       */
+/*   Updated: 2022/11/11 11:56:30 by jkollner         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,6 +21,8 @@ int	ft_printf(const char *format_string, ...);
 
 int	specifier_interpreter(char *string, va_list args);
 
-char	*hex_itoa(int n, int upper);
+char	*hex_itoa(unsigned long long n, int upper);
+
+// char	*get_addr(void *pointer);
 
 #endif /*PRINTF_H*/
